build(analyzers): adopt StyleSharp, PerformanceSharp and SecuritySharp
- Bring the .editorconfig over from the ReactiveUI repo, keeping this project's
  file header, and add a test-scoped file for the relaxations that only make
  sense in fixtures.
- Update dependencies: TUnit, Verify, NSubstitute, Splat, ReactiveUI 24 and
  System.Reactive 7. Pin the generator and analyzer to Roslyn 4.8 so they keep
  loading under VS 2022 17.8 and the .NET 8 SDK.
- Drop the unused System.Reactive reference from ReactiveUI.Binding and put the
  library on ReactiveUI.Primitives, leaving System.Reactive to
  ReactiveUI.Binding.Reactive.
- Fix a binding defect the analyzer work surfaced: the runtime stub named its
  second converter viewToVmConverter while the generated overload emitted
  viewToViewModelConverter, so a named argument bound to a different parameter
  depending on which path served the call. Both are now viewModelToViewConverter
  and viewToViewModelConverter, matching viewModelProperty.
- Seed the generator test compilations from ReactiveObject as well as
  IReactiveObject. ReactiveUI 24 moved the interface into another assembly, so
  walking references from it no longer reached the one declaring ReactiveObject
  and every ReactiveObject fixture failed to compile.
- Keep the runtime dispatch stubs as classic extension methods. As extension
  block members they outrank the generated concrete overload, so bindings fall
  through to the runtime path and throw instead of dispatching.
- Resolve the analyzer findings across the runtime library, generator, analyzer
  and tests, and regenerate the generator snapshots.

+++ src/tests/ReactiveUI.Binding.Tests/Bindings/TypeConverters/NullableByteToByteTypeConverterTests.cs
@@ -4,24 +4,16 @@
 
 namespace ReactiveUI.Binding.Tests.Bindings.TypeConverters;
 
-/// <summary>
-/// Tests for converting nullable byte to byte.
-/// </summary>
+/// <summary>Tests for converting nullable byte to byte.</summary>
 public class NullableByteToByteTypeConverterTests
 {
-    /// <summary>
-    /// Expected affinity returned for matched converter type pairs.
-    /// </summary>
+    /// <summary>Expected affinity returned for matched converter type pairs.</summary>
     private const int ExpectedAffinity = 2;
 
-    /// <summary>
-    /// Sample byte value used for conversion round-trips.
-    /// </summary>
+    /// <summary>Sample byte value used for conversion round-trips.</summary>
     private const byte SampleByte = 42;
 
-    /// <summary>
-    ///     Verifies GetAffinityForObjects Returns2.
-    /// </summary>
+    /// <summary>Verifies GetAffinityForObjects Returns2.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task GetAffinityForObjects_Returns2()
@@ -31,9 +23,7 @@ public class NullableByteToByteTypeConverterTests
         await Assert.That(affinity).IsEqualTo(ExpectedAffinity);
     }
 
-    /// <summary>
-    ///     Verifies TryConvert WithValue Succeeds.
-    /// </summary>
+    /// <summary>Verifies TryConvert WithValue Succeeds.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task TryConvert_WithValue_Succeeds()
@@ -44,12 +34,10 @@ public class NullableByteToByteTypeConverterTests
         var result = converter.TryConvert(value, null, out var output);
 
         await Assert.That(result).IsTrue();
-        await Assert.That(output).IsEqualTo((byte)SampleByte);
+        await Assert.That(output).IsEqualTo(SampleByte);
     }
 
-    /// <summary>
-    ///     Verifies TryConvert Null ReturnsFalse.
-    /// </summary>
+    /// <summary>Verifies TryConvert Null ReturnsFalse.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task TryConvert_Null_ReturnsFalse()
@@ -61,9 +49,7 @@ public class NullableByteToByteTypeConverterTests
         await Assert.That(result).IsFalse();
     }
 
-    /// <summary>
-    ///     Verifies FromType ReturnsByteNullable.
-    /// </summary>
+    /// <summary>Verifies FromType ReturnsByteNullable.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task FromType_ReturnsByteNullable()
@@ -72,9 +58,7 @@ public class NullableByteToByteTypeConverterTests
         await Assert.That(converter.FromType).IsEqualTo(typeof(byte?));
     }
 
-    /// <summary>
-    ///     Verifies ToType ReturnsByte.
-    /// </summary>
+    /// <summary>Verifies ToType ReturnsByte.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task ToType_ReturnsByte()
@@ -83,9 +67,7 @@ public class NullableByteToByteTypeConverterTests
         await Assert.That(converter.ToType).IsEqualTo(typeof(byte));
     }
 
-    /// <summary>
-    ///     Verifies TryConvertTyped WithValidValue ReturnsTrueAndOutput.
-    /// </summary>
+    /// <summary>Verifies TryConvertTyped WithValidValue ReturnsTrueAndOutput.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task TryConvertTyped_WithValidValue_ReturnsTrueAndOutput()
@@ -96,12 +78,10 @@ public class NullableByteToByteTypeConverterTests
         var success = converter.TryConvertTyped(value, null, out var result);
 
         await Assert.That(success).IsTrue();
-        await Assert.That(result).IsEqualTo((byte)SampleByte);
+        await Assert.That(result).IsEqualTo(SampleByte);
     }
 
-    /// <summary>
-    ///     Verifies TryConvertTyped WithNullValue ReturnsFalse.
-    /// </summary>
+    /// <summary>Verifies TryConvertTyped WithNullValue ReturnsFalse.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task TryConvertTyped_WithNullValue_ReturnsFalse()
@@ -114,9 +94,7 @@ public class NullableByteToByteTypeConverterTests
         await Assert.That(result).IsNull();
     }
 
-    /// <summary>
-    ///     Verifies TryConvertTyped WithInvalidType ReturnsFalse.
-    /// </summary>
+    /// <summary>Verifies TryConvertTyped WithInvalidType ReturnsFalse.</summary>
     /// <returns>A task representing the asynchronous operation.</returns>
     [Test]
     public async Task TryConvertTyped_WithInvalidType_ReturnsFalse()
